[fix]: 결제시 없는 계좌 예외처리

+++ src/main/java/com/nodaji/payment/service/PaymentService.java
@@ -3,5 +3,5 @@ package com.nodaji.payment.service;
 import java.util.Map;
 
 public interface PaymentService {
-    Map<String, Object> processPayment(String userId, String orderId, Integer amount, String paymentKey) throws Exception;
+    Map<String, Object> processPayment(String userId, String orderId, Long amount, String paymentKey) throws Exception;
 }
